flask-server/requirements.txt
cycler==0.10.0
dask==0.8.2
decorator==4.0.9
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
Pillow==3.2.0
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.4
scikit-image==0.12.3
scipy==0.17.0
six==1.10.0
toolz==0.7.4
Werkzeug==0.11.8
